product_catalog/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
aiokafka
PyYAML
pytest
